refactor(rfc64): make binding and resolver boundaries explicit
