+++ packages/chain/src/evm-context-graph-name-hash-current-slot-resolver.ts
@@ -2,7 +2,7 @@
 
 /** Bounded current-state reverse lookup over ContextGraphStorage name-hash slots. */
 
-import { ethers, type JsonRpcProvider } from 'ethers';
+import { ethers } from 'ethers';
 
 import {
   CONTEXT_GRAPH_NAME_HASH_ENUMERATION_CONCURRENCY,
@@ -10,7 +10,10 @@ import {
   type ContextGraphNameHashSlot,
   type ContextGraphNameHashSlotIndexResult,
 } from './context-graph-name-hash-resolver.js';
-import { EvmContextGraphNameHashFence } from './evm-context-graph-name-hash-fence.js';
+import type {
+  ContextGraphNameHashProviderHighWaters,
+  EvmContextGraphNameHashReader,
+} from './evm-context-graph-name-hash-fence.js';
 
 export interface EvmContextGraphNameHashCurrentSlotResolverDependencies {
   readonly onIndexCommit: () => void;
@@ -20,7 +23,7 @@ export class EvmContextGraphNameHashCurrentSlotResolver {
   private readonly slotIndex = new ContextGraphNameHashSlotIndex();
 
   constructor(
-    private readonly fence: EvmContextGraphNameHashFence,
+    private readonly reader: EvmContextGraphNameHashReader,
     private readonly dependencies: EvmContextGraphNameHashCurrentSlotResolverDependencies,
   ) {}
 
@@ -29,43 +32,32 @@ export class EvmContextGraphNameHashCurrentSlotResolver {
   }
 
   async resolve(normalizedNameHash: string): Promise<ContextGraphNameHashSlotIndexResult> {
-    await this.fence.initialize();
-    const scopeToken = await this.fence.captureScopeToken();
-    let providerHighWaters: ReadonlyMap<JsonRpcProvider, bigint> | undefined;
+    await this.reader.initialize();
+    const scopeToken = await this.reader.captureScopeToken();
     const result = await this.slotIndex.resolve(
       normalizedNameHash,
       {
-        captureScope: () => this.fence.captureScope(),
-        captureAnchor: () => this.fence.captureAnchor(),
-        loadAnchorHash: (blockNumber) => this.fence.loadAnchorHash(blockNumber),
-        loadLatestId: async () => {
-          const snapshot = await this.fence.loadProviderHighWaters();
-          providerHighWaters = snapshot.providerHighWaters;
-          return snapshot.latestId;
-        },
-        loadRange: (firstId, lastId) => {
-          if (providerHighWaters === undefined) {
-            throw new Error(
-              'resolveContextGraphIdByNameHash: current provider high-water snapshot is missing',
-            );
-          }
-          return this.loadSlots(firstId, lastId, providerHighWaters);
-        },
+        captureScope: () => this.reader.captureScope(),
+        captureAnchor: () => this.reader.captureAnchor(),
+        loadAnchorHash: (blockNumber) => this.reader.loadAnchorHash(blockNumber),
+        loadHighWaterSnapshot: () => this.reader.loadProviderHighWaters(),
+        loadRange: (firstId, lastId, snapshot) =>
+          this.loadSlots(firstId, lastId, snapshot),
         onCommit: this.dependencies.onIndexCommit,
       },
     );
     if (result.mode === 'historical') return result;
 
+    const verification = await this.reader.loadProviderHighWaters();
+    if (verification.latestId !== result.highWater) {
+      throw new Error(
+        `resolveContextGraphIdByNameHash: Context Graph registry advanced from ` +
+        `${result.highWater.toString()} to ${verification.latestId.toString()} ` +
+        'during current-slot resolution',
+      );
+    }
     if (result.id !== null) {
-      const verification = await this.fence.loadProviderHighWaters();
-      if (verification.latestId !== result.highWater) {
-        throw new Error(
-          `resolveContextGraphIdByNameHash: Context Graph registry advanced from ` +
-          `${result.highWater.toString()} to ${verification.latestId.toString()} ` +
-          'during current-slot resolution',
-        );
-      }
-      const currentHash = await this.fence.readCurrentNameHash(
+      const currentHash = await this.reader.readCurrentNameHash(
         result.id,
         undefined,
         verification.providerHighWaters,
@@ -79,7 +71,7 @@ export class EvmContextGraphNameHashCurrentSlotResolver {
       }
     }
 
-    await this.fence.assertScopeCurrent(scopeToken, 'current-slot resolution');
+    await this.reader.assertScopeCurrent(scopeToken, 'current-slot resolution');
     return result;
   }
 
@@ -87,7 +79,7 @@ export class EvmContextGraphNameHashCurrentSlotResolver {
   private async loadSlots(
     firstId: bigint,
     lastId: bigint,
-    providerHighWaters: ReadonlyMap<JsonRpcProvider, bigint>,
+    highWaterSnapshot: ContextGraphNameHashProviderHighWaters,
   ): Promise<readonly ContextGraphNameHashSlot[]> {
     const scanController = new AbortController();
     const slots: ContextGraphNameHashSlot[] = [];
@@ -100,10 +92,10 @@ export class EvmContextGraphNameHashCurrentSlotResolver {
         if (contextGraphId > lastId) return;
         nextId += 1n;
         try {
-          const currentHash = await this.fence.readCurrentNameHash(
+          const currentHash = await this.reader.readCurrentNameHash(
             contextGraphId,
             scanController.signal,
-            providerHighWaters,
+            highWaterSnapshot.providerHighWaters,
           );
           slots.push({ id: contextGraphId, nameHash: currentHash });
         } catch (cause) {
