+++ packages/chain/src/evm-context-graph-name-hash-historical-log-resolver.ts
@@ -2,62 +2,21 @@
 
 /** Bounded deploy-anchored exact-topic ContextGraphCreated reverse lookup. */
 
-import { Contract, ethers, type JsonRpcProvider } from 'ethers';
+import { ethers, type JsonRpcProvider } from 'ethers';
 
 import { CG_REGISTRY_MAX_SCAN_PAGES } from './evm-adapter-constants.js';
-import { EvmContextGraphNameHashFence } from './evm-context-graph-name-hash-fence.js';
-
-export interface ContextGraphNameHashScanProvider {
-  readonly provider: JsonRpcProvider;
-  readonly backendHead: number;
-}
-
-export interface EvmContextGraphNameHashHistoricalLogResolverDependencies {
-  readonly requireContextGraphStorage: () => Contract;
-  readonly scanPageSize: () => number;
-  readonly resolveContractDeployBlock: (
-    address: string,
-    operationLabel: string,
-    contractLabel: string,
-  ) => Promise<{
-    readonly fromBlock: number;
-    readonly head: number;
-    readonly scanProviders: ReadonlyArray<ContextGraphNameHashScanProvider>;
-    readonly degradedFromGenesis?: boolean;
-  }>;
-  readonly queryEventLogsPage: (
-    baseContract: Contract,
-    filter: unknown,
-    lo: number,
-    hi: number,
-    scanProviders: ReadonlyArray<ContextGraphNameHashScanProvider>,
-    connected: Map<JsonRpcProvider, Contract>,
-    label: string,
-    preferred?: JsonRpcProvider,
-  ) => Promise<{
-    readonly logs: ReadonlyArray<ethers.EventLog | ethers.Log>;
-    readonly provider: JsonRpcProvider;
-  }>;
-}
+import type { EvmContextGraphNameHashReader } from './evm-context-graph-name-hash-fence.js';
 
 export class EvmContextGraphNameHashHistoricalLogResolver {
   constructor(
-    private readonly fence: EvmContextGraphNameHashFence,
-    private readonly dependencies: EvmContextGraphNameHashHistoricalLogResolverDependencies,
+    private readonly reader: EvmContextGraphNameHashReader,
   ) {}
 
   async resolve(normalizedNameHash: string): Promise<bigint | null> {
-    await this.fence.initialize();
-    const scopeToken = await this.fence.captureScopeToken();
-    const cgs = this.dependencies.requireContextGraphStorage();
-    const storageAddress = (await cgs.getAddress()).toLowerCase();
-    const { fromBlock, head, scanProviders: reachableProviders } =
-      await this.dependencies.resolveContractDeployBlock(
-        storageAddress,
-        'resolveContextGraphIdByNameHash',
-        'ContextGraphStorage',
-      );
-    const pageSize = this.dependencies.scanPageSize();
+    await this.reader.initialize();
+    const scopeToken = await this.reader.captureScopeToken();
+    const scan = await this.reader.prepareHistoricalScan();
+    const { fromBlock, head, pageSize } = scan;
     const pages = fromBlock > head
       ? 0
       : Math.ceil((head - fromBlock + 1) / pageSize);
@@ -70,13 +29,14 @@ export class EvmContextGraphNameHashHistoricalLogResolver {
       );
     }
 
-    const headAnchor = await this.fence.captureHistoricalHead(reachableProviders, head);
-    const scannedRegistryHighWater = await this.fence.loadHistoricalRegistryHighWaterAtHead(
+    const anchoredScan = await scan.anchor();
+    const { headAnchor } = anchoredScan;
+    const scannedRegistryHighWater = await this.reader.loadHistoricalRegistryHighWaterAtHead(
       headAnchor.scanProviders.map(({ provider }) => provider),
       head,
     );
     const assertHistoricalRegistryCurrent = async (): Promise<void> => {
-      const currentBoundary = await this.fence.loadProviderHighWaters();
+      const currentBoundary = await this.reader.loadProviderHighWaters();
       if (currentBoundary.latestId !== scannedRegistryHighWater) {
         throw new Error(
           `resolveContextGraphIdByNameHash: registry high-water changed from ` +
@@ -90,8 +50,8 @@ export class EvmContextGraphNameHashHistoricalLogResolver {
       headAnchor.scanProviders[0]!.provider,
     ]);
     const assertScanCurrent = async (): Promise<void> => {
-      await this.fence.assertScopeCurrent(scopeToken, 'historical scan');
-      await this.fence.assertHistoricalHeadCurrent(headAnchor, usedProviders);
+      await this.reader.assertScopeCurrent(scopeToken, 'historical scan');
+      await this.reader.assertHistoricalHeadCurrent(headAnchor, usedProviders);
     };
 
     if (fromBlock > head) {
@@ -101,28 +61,19 @@ export class EvmContextGraphNameHashHistoricalLogResolver {
       return null;
     }
 
-    const filter = cgs.filters.ContextGraphCreated(null, null, normalizedNameHash);
-    const connected = new Map<JsonRpcProvider, Contract>();
     const ids = new Set<bigint>();
     let preferred: JsonRpcProvider | undefined;
     for (let lo = fromBlock; lo <= head; lo += pageSize) {
       const hi = Math.min(lo + pageSize - 1, head);
-      const page = await this.dependencies.queryEventLogsPage(
-        cgs,
-        filter,
+      const page = await anchoredScan.readContextGraphCreatedPage(
+        normalizedNameHash,
         lo,
         hi,
-        headAnchor.scanProviders,
-        connected,
-        'resolveContextGraphIdByNameHash ContextGraphCreated',
         preferred,
       );
       preferred = page.provider;
       usedProviders.add(page.provider);
-      for (const log of page.logs) {
-        const parsed = cgs.interface.parseLog({ topics: [...log.topics], data: log.data });
-        if (parsed?.name !== 'ContextGraphCreated') continue;
-        const id = BigInt(parsed.args.contextGraphId);
+      for (const id of page.ids) {
         if (id <= 0n) {
           throw new Error(
             `resolveContextGraphIdByNameHash: invalid Context Graph id ` +
@@ -147,7 +98,7 @@ export class EvmContextGraphNameHashHistoricalLogResolver {
     }
 
     const id = ids.values().next().value as bigint;
-    const currentHash = await this.fence.readCurrentNameHash(id);
+    const currentHash = await this.reader.readCurrentNameHash(id);
     if (currentHash !== normalizedNameHash) {
       throw new Error(
         `resolveContextGraphIdByNameHash: slot ${id.toString()} currently commits ` +
